+++ src/pages/demoapp.tsx
@@ -1,15 +1,15 @@
 import * as React from "react";
 import styled from "styled-components/macro";
-import UnderArbeid from "../templates/veiviser/UnderArbeid";
-import VelgDetSomPasserBest, { VeiviserValg } from "../templates/veiviser/VelgDetSomPasserBest";
-import { veiviserMachine } from "../templates/veiviser/VeiviserStateChart";
+import UnderArbeid from "../components/veiviser/UnderArbeid";
+import VelgDetSomPasserBest, { VeiviserValg } from "../components/veiviser/VelgDetSomPasserBest";
+import { veiviserMachine } from "../components/veiviser/VeiviserStateChart";
 import { useMachine } from "@xstate/react";
 import getAlleTilpassInnholdValg from "../components/faktaside/TilpassInnhold/getAlleTilpassInnholdValg";
 import { Group, isGroup } from "../utils/richTextUtils/richTextTypes";
 import BlockContent from "../components/BlockContent/BlockContent";
 import { visBasertPåFiltrering } from "../components/BlockContent/VisFor/VisFor";
 import { typografiStyle } from "../components/faktaside/MainContentStyle";
-import VeiviserBrødsmuler from "../templates/veiviser/VeiviserBrødsmuler";
+import VeiviserBrødsmuler from "../components/veiviser/VeiviserBrødsmuler";
 import DevKnapper from "../components/DevKnapper/DevKnapper";
 import { useVisForContext } from "../components/BlockContent/VisFor/VisForContext";
 import { visBasertPaaVisPaaConfig } from "../components/BlockContent/VisFor/VisPaaSide";
